+++ ds-backend/src/routes/payment.routes.js
@@ -1,7 +1,7 @@
 import { Router } from "express";
 import { makePayment, updatePayment } from "../controllers/payment.controller";
 const paymentRouter = Router();
-import { doubleCsrfProtection, csrfErrorHandler } from "../utils/csrf";
+import { doubleCsrfProtection, csrfErrorHandler } from "../middleware/csrf";
 
 paymentRouter.post("/", doubleCsrfProtection, csrfErrorHandler,makePayment);
 paymentRouter.patch(
