+++ ds-frontend/.gitignore
@@ -27,3 +27,4 @@ dist-ssr
 .env
 
 .scannerwork/
+scripts
